added some new depency and added glance section for landing page
