+++ package.json
@@ -22,9 +22,11 @@
     "framer-motion": "^10.0.0",
     "next": "^15.0.3",
     "react": "^18.2.0",
+    "react-countup": "^6.5.3",
     "react-datepicker": "^7.5.0",
     "react-dom": "^18.2.0",
-    "react-icons": "^4.7.1"
+    "react-icons": "^4.7.1",
+    "react-intersection-observer": "^9.16.0"
   },
   "devDependencies": {
     "@types/node": "18.8.4",
